[refactor] split SecondFactorEditDialog into Model and View
