+++ src/misc/2fa/webauthn/WebauthnClient.ts
@@ -160,11 +160,9 @@ const WEBAUTHN_STRING_MAX_BYTE_LENGTH = 64
  * some authenticators truncate this and others refuse to be registered
  * at all if this validation does not pass.
  *
- * technically, we'd also be supposed to encode text direction and a language
+ * Note: technically, we'd also be supposed to encode text direction and a language
  * code into the display name.
  */
 export function validateWebauthnDisplayName(displayName: string): boolean {
-	const ret = WEBAUTHN_STRING_MAX_BYTE_LENGTH - stringToUtf8Uint8Array(displayName).byteLength >= 0
-	console.log('ret=', ret)
-	return ret
+	return WEBAUTHN_STRING_MAX_BYTE_LENGTH - stringToUtf8Uint8Array(displayName).byteLength >= 0
 }
